# docs/rest/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Documentation for the Project Odeon backend",
    "version" : "v2",
    "title" : "Swagger documentation for the Project Odeon backend"
  },
  "host" : "jscpeterson.com",
  "basePath" : "/rest/",
  "schemes" : [ "https" ],
  "paths" : {
...
    "/film_locations/{filmLocationId}/images/{imageId}" : {
      "get" : {
        "summary" : "Gets all images on a FilmLocation.",
        "description" : "",
        "operationId" : "getImage",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "filmLocationId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "imageId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Image"
            }
          }
        }
      },
      "delete" : {
        "summary" : "Deletes an image on a film location.",
        "description" : "",
        "operationId" : "deleteImage",
        "parameters" : [ {
          "name" : "filmLocationId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "imageId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "204" : {
            "description" : ""
          }
        }
      }
    },
...
    "/images" : {
      "get" : {
        "summary" : "Orders Images",
        "description" : "Orders by Id ascending.",
        "operationId" : "list",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Image"
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "Post Image",
        "description" : "Saves image to image repository",
        "operationId" : "post",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Image"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Image"
            }
          }
        }
      }
    },
    "/images/{imageId}" : {
      "get" : {
        "summary" : "Gets an image",
        "description" : "Finds image by the image UUID",
        "operationId" : "get",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "imageId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Image"
            }
          }
        }
      },
      "delete" : {
        "summary" : "Deletes an image",
        "description" : "Enables you to delete an image from database",
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "imageId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "204" : {
            "description" : ""
          }
        }
      }
    },
    "/productions" : {
      "get" : {
        "operationId" : "list",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Production"
              }
            }
          }
        }
      },
      "post" : {
        "operationId" : "post",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Production"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Production"
            }
          }
        }
      },
      "patch" : {
        "operationId" : "patch",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Production"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/productions/{productionId}" : {
      "get" : {
        "operationId" : "get",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "productionId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Production"
            }
          }
        }
      },
      "delete" : {
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "productionId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "204" : {
            "description" : ""
          }
        }
      }
    },
    "/productions/{productionId}/poster" : {
      "get" : {
        "operationId" : "getPoster",
        "produces" : [ "image/gif", "image/jpeg", "image/png", "video/webm", "video/mp4", "application/octet-stream" ],
        "parameters" : [ {
          "name" : "productionId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/StreamingResponseBody"
            }
          }
        }
      }
    },
...
    "/users" : {
      "get" : {
        "summary" : "Gets all users",
        "description" : "Retrieves all users with their ID in an list.",
        "operationId" : "list",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/GoogleUser"
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "Posts a google user.",
        "description" : "Posts a googel user.",
        "operationId" : "post",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/GoogleUser"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/GoogleUser"
            }
          }
        }
      },
      "patch" : {
        "summary" : "Patches a google user",
        "description" : "Patches a user. This will overwrite everything. if just changing one field all other current fields must be included.",
        "operationId" : "patch",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/GoogleUser"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/users/{userId}" : {
      "get" : {
        "summary" : "Gets google user",
        "description" : "Gets google user id and returns google user.",
        "operationId" : "get",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/GoogleUser"
            }
          }
        }
      },
      "delete" : {
        "summary" : "Deletes google user",
        "description" : "Takes user id as a parameter and deletes google user.",
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "204" : {
            "description" : ""
          }
        }
      }
    }
  },
  "definitions" : {
    "FilmLocation" : {
      "type" : "object",
      "properties" : {
        "userId" : {
          "type" : "string",
          "format" : "uuid"
        },
        "user" : {
          "$ref" : "#/definitions/GoogleUser"
        },
        "id" : {
          "type" : "string",
          "format" : "uuid"
        },
        "longCoordinate" : {
          "type" : "number",
          "format" : "double"
        },
        "latCoordinate" : {
          "type" : "number",
          "format" : "double"
        },
        "siteName" : {
          "type" : "string"
        },
        "imdbId" : {
          "type" : "string"
        },
        "productionId" : {
          "type" : "string"
        },
        "production" : {
          "$ref" : "#/definitions/Production"
        },
        "address" : {
          "type" : "string"
        },
        "shootDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "originalDetails" : {
          "type" : "string"
        },
        "approved" : {
          "type" : "boolean"
        },
        "googleId" : {
          "type" : "string"
        },
        "href" : {
          "type" : "string",
          "format" : "uri"
        }
      }
    },
    "GoogleUser" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid"
        },
        "googleId" : {
          "type" : "string"
        },
        "googleName" : {
          "type" : "string"
        },
        "gmailAddress" : {
          "type" : "string"
        },
        "banned" : {
          "type" : "boolean"
        },
        "banReason" : {
          "type" : "string"
        },
        "userRole" : {
          "type" : "string"
        },
        "href" : {
          "type" : "string",
          "format" : "uri"
        }
      }
    },
    "Image" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid"
        },
        "filmLocation" : {
          "$ref" : "#/definitions/FilmLocation"
        },
        "description" : {
          "type" : "string"
        },
        "url" : {
          "type" : "string"
        },
        "userId" : {
          "type" : "string",
          "format" : "uuid"
        },
        "user" : {
          "$ref" : "#/definitions/GoogleUser"
        },
        "googleId" : {
          "type" : "string"
        },
        "approved" : {
          "type" : "boolean"
        },
        "href" : {
          "type" : "string",
          "format" : "uri"
        }
      }
    },
    "Production" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid"
        },
        "imdbId" : {
          "type" : "string"
        },
        "title" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        },
        "releaseYear" : {
          "type" : "string"
        },
        "plot" : {
          "type" : "string"
        },
        "href" : {
          "type" : "string",
          "format" : "uri"
        }
      }
    },
    "StreamingResponseBody" : {
      "type" : "object"
    },
...
  }
}